many errors fixed

+++ src/distance/distance-ros-1.py
@@ -3,7 +3,7 @@ import rospy
 import distance
 import distance_pf
 from geometry_msgs.msg import Twist
-from std_msgs.msg import Float64
+from std_msgs.msg import Float32
 from nav_msgs.msg import Odometry
 from bearnav2.srv import SetDist, SetDistResponse
 from bearnav2.msg import PFInput
@@ -39,7 +39,7 @@ if __name__ == "__main__":
     odom_topic = rospy.get_param("~odom_topic")
     # d = distance.Distance(use_twist)
     d = distance_pf.DistancePF(False)
-    pub = rospy.Publisher("distance", Float64, queue_size=1)
+    pub = rospy.Publisher("distance", Float32, queue_size=1)
     rospy.Subscriber(odom_topic, Odometry, callbackOdom, queue_size=1)
     rospy.Subscriber(cmd_vel_topic, Twist, callbackTwist, queue_size=1)
     rospy.Subscriber("pf_img_input", PFInput, callbackCamera)  # TODO: make name of topic as argument
